split core/utility.hpp into core/logger.hpp and core/utility/singleton.hpp
