+++ SeQuant/core/utility/singleton.hpp
@@ -2,10 +2,11 @@
 // Created by Eduard Valeyev on 2019-03-11.
 //
 
-#ifndef SEQUANT_UTILITY_HPP
-#define SEQUANT_UTILITY_HPP
+#ifndef SEQUANT_CORE_UTILITY_SINGLETON_HPP
+#define SEQUANT_CORE_UTILITY_SINGLETON_HPP
 
 #include <memory>
+#include <type_traits>
 
 namespace sequant {
 
@@ -89,34 +90,6 @@ class Singleton {
   }
 };
 
-struct Logger : public Singleton<Logger> {
-  bool wick_harness = false;
-  bool wick_topology = false;
-  bool wick_contract = false;
-  bool wick_reduce = false;
-  bool wick_stats = false;
-  bool expand = false;
-  bool canonicalize = false;
-  bool canonicalize_dot = false;
-  bool simplify = false;
-  bool tensor_network = false;
-
- private:
-  friend class Singleton<Logger>;
-  Logger(int log_level = 0) {
-    if (log_level > 0) {
-      wick_topology = true;
-      wick_contract = true;
-      wick_reduce = true;
-      wick_stats = true;
-      expand = true;
-      canonicalize = true;
-      simplify = true;
-      tensor_network = true;
-    }
-  }
-};
-
 }  // namespace sequant
 
-#endif  // SEQUANT_UTILITY_HPP
+#endif  // SEQUANT_CORE_UTILITY_SINGLETON_HPP
