+++ tests/unit/test_main.cpp
@@ -4,10 +4,10 @@
 
 #define CATCH_CONFIG_RUNNER
 #include <clocale>
+#include "SeQuant/core/logger.hpp"
 #include "SeQuant/core/op.hpp"
 #include "SeQuant/core/runtime.hpp"
 #include "SeQuant/core/space.hpp"
-#include "SeQuant/core/utility.hpp"
 #include "SeQuant/domain/mbpt/convention.hpp"
 #include "SeQuant/domain/mbpt/formalism.hpp"
 #include "catch.hpp"
